feat: move singleton class into separate file to avoid circular imports.
