+++ yearn/prices/uniswap/v2.py
@@ -13,7 +13,8 @@ from yearn.multicall2 import fetch_multicall
 from yearn.networks import Network
 from yearn.prices.constants import stablecoins, usdc, weth
 from yearn.typing import Address, AddressOrContract, Block
-from yearn.utils import Singleton, contract
+from yearn.utils import contract
+from yearn.singleton import Singleton
 
 logger = logging.getLogger(__name__)
