+++ yearn/v2/registry.py
@@ -10,7 +10,8 @@ from web3._utils.events import construct_event_topic_set
 from yearn.events import create_filter, decode_logs, get_logs_asap
 from yearn.multicall2 import fetch_multicall
 from yearn.prices import magic
-from yearn.utils import Singleton, contract_creation_block, contract
+from yearn.utils import contract_creation_block, contract, get_start_block
+from yearn.singleton import Singleton
 from yearn.v2.vaults import Vault
 from yearn.networks import Network
 from yearn.exceptions import UnsupportedNetwork
